+++ Audibly.App/Views/ControlPages/ChangelogDialogContent.xaml.cs
@@ -1,7 +1,8 @@
 // Author: rstewa · https://github.com/rstewa
-// Created: 4/11/2024
-// Updated: 4/11/2024
+// Created: 4/15/2024
+// Updated: 6/1/2024
 
+using Audibly.App.Helpers;
 using Microsoft.UI.Xaml.Controls;
 
 namespace Audibly.App.Views.ControlPages;
@@ -14,10 +15,13 @@ public sealed partial class ChangelogDialogContent : Page
     public string Title { get; set; }
 
     public string ChangelogText { get; set; }
-    
-    public ChangelogDialogContent(string title, string changelogText)
+
+    public string Subtitle { get; set; }
+
+    public ChangelogDialogContent(string changelogText)
     {
-        Title = title;
+        Title = "What's New?";
+        Subtitle = $"Version {Constants.Version}";
         ChangelogText = changelogText;
         InitializeComponent();
     }