+++ Audibly.App/Views/SettingsPage.xaml.cs
@@ -20,21 +20,7 @@ public sealed partial class SettingsPage : Page
     /// </summary>
     public MainViewModel ViewModel => App.ViewModel;
 
-    public string Version
-    {
-        get
-        {
-            var assembly = Assembly.GetEntryAssembly();
-            if (assembly == null) 
-                return string.Empty;
-
-            var version = assembly.GetName().Version;
-            if (version == null) 
-                return string.Empty;
-
-            return string.Format("{0}.{1}.{2}.{3}", version.Major, version.Minor, version.Build, version.Revision);
-        }
-    }
+    public static string Version => Constants.Version;
 
     public SettingsPage()
     {