+++ CHANGELOG.md
@@ -9,7 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added file activation support (for opening audiobooks from the file explorer and setting Audibly as the default app for opening audiobooks)
+- Added file activation support
+  - Users can now open audiobooks (.m4b) from the file explorer
+  - Users can also set Audibly as the default app for opening audiobooks (.m4b)
+
+- Added Changelog dialog for new versions
+  - Users will now see a dialog with the changelog when they update the app
 
 ### Fixed